# openapi.yaml
openapi: 3.0.3
info:
  title: EasyBooking API
  version: 1.0.0
  description: REST API for the EasyBooking endterm project.
servers:
  - url: http://127.0.0.1:3000
paths:
...
  /api/hotels:
    get:
      summary: List hotels
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1 }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string, example: rating_desc }
      responses:
        '200':
          description: Hotels list
    post:
      summary: Create hotel (admin)
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotelCreateInput'
      responses:
        '201':
          description: Created
        '403':
          description: Forbidden
...
  /api/bookings:
    get:
      summary: List bookings (auth)
      security: [{ cookieAuth: [] }]
      parameters:
        - in: query
          name: scope
          schema:
            type: string
            enum: [mine, all]
        - in: query
          name: status
          schema:
            type: string
            enum: [confirmed, cancelled]
      responses:
        '200':
          description: Bookings list
    post:
      summary: Create booking (auth)
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateInput'
      responses:
        '201':
          description: Created
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: easybook.sid
...
  schemas:
...
    HotelCreateInput:
      type: object
      required:
        - title
        - description
        - location
        - address
        - price_per_night
        - rating
        - available_rooms
        - amenities
      properties:
        title: { type: string }
        description: { type: string }
        location: { type: string }
        address: { type: string }
        price_per_night: { type: number }
        rating: { type: number }
        available_rooms: { type: integer }
        amenities:
          type: array
          items: { type: string }
        imageUrl: { type: string }
...
    BookingCreateInput:
      type: object
      required: [hotelId, checkIn, checkOut, guests]
      properties:
        hotelId: { type: string }
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }
        guests: { type: integer, minimum: 1, maximum: 10 }
        status: { type: string, enum: [confirmed, cancelled] }
        notes: { type: string }
...